scraper/requirements.txt
playwright
openai
beautifulsoup4
python-dotenv
supabase
